RPC, DISPATCHER: Integrated dispatcher into perun-rpc
- Modified integration of dispatcher into RPC.
- Fixed loading log4j config. It must read log4j.xml, not log4j.properties file.
- Removed dispatcher appcontext and DispatcherStarter from RPC.
  Moved to the perun-dispatcher module (overwrite original DispatcherStarter).
  Updated appcontex accordingly.
- Fixed session user for dispatcher in appcontext config.
- Removed loading jdbc.properties from propertiesBean, since it's not read from
  there.
- Added dynamic property placeholder for jdbc.properties instead of fixed /etc/perun.
- Source format of xml files.
