+++ perun-dispatcher-new/src/main/java/cz/metacentrum/perun/dispatcher/main/DispatcherStarter.java
@@ -1,11 +1,7 @@
 package cz.metacentrum.perun.dispatcher.main;
 
-import java.awt.HeadlessException;
-import java.awt.SplashScreen;
-
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.springframework.context.support.ClassPathXmlApplicationContext;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.support.AbstractApplicationContext;
@@ -20,54 +16,43 @@ import cz.metacentrum.perun.dispatcher.service.DispatcherManager;
  */
 public class DispatcherStarter {
 
-	private final static Logger log = LoggerFactory
-			.getLogger(DispatcherStarter.class);
+	private final static Logger log = LoggerFactory.getLogger(DispatcherStarter.class);
 
 	private DispatcherManager dispatcherManager;
+	@Autowired
 	private AbstractApplicationContext springCtx;
 
-	public DispatcherStarter() {
-		springCtx = new ClassPathXmlApplicationContext(
-				"/perun-dispatcher-applicationcontext.xml",
-				"/perun-dispatcher-applicationcontext-jdbc.xml");
-		this.dispatcherManager = springCtx.getBean("dispatcherManager",
-				DispatcherManager.class);
-	}
-
-	public static void main(String[] args) {
+	/**
+	 * Initialize integrated dispatcher.
+	 */
+	public final void init() {
 		try {
 
-			DispatcherStarter dispatcherStarter = new DispatcherStarter();
-			// Just for the Spring IoC to exit gracefully...
-			dispatcherStarter.springCtx.registerShutdownHook();
+			dispatcherManager = springCtx.getBean("dispatcherManager", DispatcherManager.class);
+			springCtx.registerShutdownHook();
 
 			// Register into the database
 			// DO NOT: dispatcherStarter.dispatcherManager.registerDispatcher();
 			// Start HornetQ server
-			dispatcherStarter.dispatcherManager.startPerunHornetQServer();
+			dispatcherManager.startPerunHornetQServer();
 			// Start System Queue Processor
-			dispatcherStarter.dispatcherManager.startProcessingSystemMessages();
-			// Prefetch rules for all the Engnes in the Perun DB and create
+			dispatcherManager.startProcessingSystemMessages();
+			// Prefetch rules for all the Engines in the Perun DB and create
 			// Dispatcher queues
-			dispatcherStarter.dispatcherManager
-					.prefetchRulesAndDispatcherQueues();
-			// reload tasksk from database
-			dispatcherStarter.dispatcherManager.loadSchedulingPool();
+			dispatcherManager.prefetchRulesAndDispatcherQueues();
+			// reload tasks from database
+			dispatcherManager.loadSchedulingPool();
 			// Start parsers (mining data both from Grouper and PerunDB)
-			dispatcherStarter.dispatcherManager.startParsingData();
+			dispatcherManager.startParsingData();
 			// Start Event Processor
-			dispatcherStarter.dispatcherManager.startPocessingEvents();
+			dispatcherManager.startPocessingEvents();
 			log.info("Done. Perun-Dispatcher has started.");
-			System.out.println("Done. Perun-Dispatcher has started.");
-			if (SplashScreen.getSplashScreen() != null) {
-				SplashScreen.getSplashScreen().close();
-			}
-		} catch (HeadlessException e) {
-			// Doesn't matter... (We can't show splash screen on a server :-))
 		} catch (PerunHornetQServerException e) {
 			log.error(e.toString(), e);
 		} catch (Exception e) {
 			log.error(e.toString(), e);
 		}
+
 	}
+
 }
