+++ perun-dispatcher-new/src/main/resources/perun-dispatcher-applicationcontext-jdbc.xml
@@ -11,33 +11,52 @@ http://www.springframework.org/schema/aop http://www.springframework.org/schema/
 http://www.springframework.org/schema/tx http://www.springframework.org/schema/tx/spring-tx.xsd
 ">
 
-    <context:property-placeholder location="file:/etc/perun/jdbc.properties"/>
-    
-    	<!-- DataSource implementation -->
-    <bean id="dataSource" class="org.apache.tomcat.dbcp.dbcp.BasicDataSource" destroy-method="close">
-        <property name="driverClassName" value="${jdbc.driver}"/>
-        <property name="url" value="${jdbc.url}"/>
-        <property name="username" value="${jdbc.username}"/>
-        <property name="password" value="${jdbc.password}"/>
-        <property name="maxActive" value="10"/>
-    </bean>
 
-    <!-- Transaction manager for Spring JDBC -->
-    <bean id="transactionManager" class="org.springframework.jdbc.datasource.DataSourceTransactionManager">
-        <property name="dataSource" ref="dataSource"/>
-    </bean>
+	<!-- Set property placeholder for JDBC dataSource -->
+	<bean class="org.springframework.beans.factory.config.PropertyPlaceholderConfigurer">
+		<property name="ignoreResourceNotFound" value="true"/>
+		<property name="ignoreUnresolvablePlaceholders" value="true"/>
+		<property name="locations">
+			<list>
+				<value>${perun.jdbc}</value>
+				<value>file:${perun.conf.custom}jdbc.properties</value>
+			</list>
+		</property>
+		<property name="properties">
+			<props>
+				<prop key="jdbc.driver"></prop>
+				<prop key="jdbc.url"></prop>
+				<prop key="jdbc.username"></prop>
+				<prop key="jdbc.password"></prop>
+			</props>
+		</property>
+	</bean>
 
-    <!-- DAOs -->
-    <bean id="dispatcherDao" class="cz.metacentrum.perun.dispatcher.dao.jdbc.DispatcherDaoJdbc">
-    	<property name="dataSource" ref="dataSource"/>
-    </bean>
+	<!-- DataSource implementation -->
+	<bean id="dataSource" class="org.apache.tomcat.dbcp.dbcp.BasicDataSource" destroy-method="close">
+		<property name="driverClassName" value="${jdbc.driver}"/>
+		<property name="url" value="${jdbc.url}"/>
+		<property name="username" value="${jdbc.username}"/>
+		<property name="password" value="${jdbc.password}"/>
+		<property name="maxActive" value="10"/>
+	</bean>
 
-    <bean id="rulesDao" class="cz.metacentrum.perun.dispatcher.dao.jdbc.RulesDaoJdbc">
-    	<property name="dataSource" ref="dataSource"/>
-    </bean>
+	<!-- Transaction manager for Spring JDBC -->
+	<bean id="transactionManager" class="org.springframework.jdbc.datasource.DataSourceTransactionManager">
+		<property name="dataSource" ref="dataSource"/>
+	</bean>
 
-    <bean id="processingRuleDao" class="cz.metacentrum.perun.dispatcher.dao.jdbc.ProcessingRuleDaoJdbc">
-      <property name="dataSource" ref="dataSource"/> 
-    </bean>
+	<!-- DAOs -->
+	<bean id="dispatcherDao" class="cz.metacentrum.perun.dispatcher.dao.jdbc.DispatcherDaoJdbc">
+		<property name="dataSource" ref="dataSource"/>
+	</bean>
+
+	<bean id="rulesDao" class="cz.metacentrum.perun.dispatcher.dao.jdbc.RulesDaoJdbc">
+		<property name="dataSource" ref="dataSource"/>
+	</bean>
+
+	<bean id="processingRuleDao" class="cz.metacentrum.perun.dispatcher.dao.jdbc.ProcessingRuleDaoJdbc">
+		<property name="dataSource" ref="dataSource"/>
+	</bean>
 
 </beans>
