+++ perun-dispatcher-new/src/main/resources/perun-dispatcher-applicationcontext.xml
@@ -13,64 +13,64 @@ http://www.springframework.org/schema/tx http://www.springframework.org/schema/t
 http://www.springframework.org/schema/task http://www.springframework.org/schema/task/spring-task.xsd
 ">
 
-    <!-- Enable @Transactional support -->
-    <tx:annotation-driven/>
+	<!-- Enable @Transactional support -->
+	<tx:annotation-driven/>
 
-    <!-- Enable @AspectJ support -->
-    <aop:aspectj-autoproxy/>
+	<!-- Enable @AspectJ support -->
+	<aop:aspectj-autoproxy/>
 
-    <!--It basically says, wrap any interface in the service package that ends
+	<!--It basically says, wrap any interface in the service package that ends
 	with "Manager" with a transaction. Because of this, it's important to
 	set your dependencies to interfaces, not to implementations.
-    <aop:config>
-        <!<aop:advisor id="managerTx" advice-ref="txAdvice" pointcut="execution(* *..service.*Manager.*(..))"/>
-    </aop:config>-->
+	<aop:config>
+		<!<aop:advisor id="managerTx" advice-ref="txAdvice" pointcut="execution(* *..service.*Manager.*(..))"/>
+	</aop:config>-->
 
-    <!-- Importing Perun Core... -->
-    <import resource="classpath:perun-beans.xml"/>
+	<!-- Importing Perun Core... -->
+	<import resource="classpath:perun-beans.xml"/>
 	<!-- Import Perun taskslib -->
-    <import resource="classpath:perun-tasks-lib-applicationcontext.xml"/>
-    	
-    <tx:advice id="txAdvice">
-        <tx:attributes>
-            <tx:method name="get*" read-only="true"/>
-            <tx:method name="*"/>
-        </tx:attributes>
-    </tx:advice>
+	<import resource="classpath:perun-tasks-lib-applicationcontext.xml"/>
 
-    <!-- Scans for @Repository, @Service and @Component -->
-    <context:component-scan base-package="cz.metacentrum.perun.dispatcher"/>
-    <context:component-scan base-package="cz.metacentrum.perun.controller"/>
-    
+	<tx:advice id="txAdvice">
+		<tx:attributes>
+			<tx:method name="get*" read-only="true"/>
+			<tx:method name="*"/>
+		</tx:attributes>
+	</tx:advice>
+
+	<!-- Scans for @Repository, @Service and @Component -->
+	<context:component-scan base-package="cz.metacentrum.perun.dispatcher"/>
+	<context:component-scan base-package="cz.metacentrum.perun.controller"/>
+
+	<bean id="dispatcherStarter" class="cz.metacentrum.perun.dispatcher.main.DispatcherStarter" init-method="init" />
+
 	<bean id="taskExecutor" class="org.springframework.scheduling.concurrent.ThreadPoolTaskExecutor">
-  		<property name="corePoolSize" value="10" />
-  		<property name="maxPoolSize" value="50" />
-  		<property name="queueCapacity" value="100" />
+		<property name="corePoolSize" value="10" />
+		<property name="maxPoolSize" value="50" />
+		<property name="queueCapacity" value="100" />
 	</bean>
 
-    <!-- Properties Bean -->
-    <bean id="dispatcherPropertiesBean" class="org.springframework.beans.factory.config.PropertiesFactoryBean">
-        <property name="locations">
-        	<list>
-	           <value>${perun.jdbc}</value>
-		   <value>file:${perun.conf.custom}jdbc.properties</value>
-	           <value>file:${perun.conf}perun-dispatcher.properties</value>
-                   <value>file:${perun.conf.custom}perun-dispatcher.properties</value>
-                </list>
-        </property>
-        <property name="ignoreResourceNotFound">
-            <value>true</value>
-        </property>
-       	<property name="properties">
-       		<props>
-       		    	<prop key="perun.principal.name">perunTests</prop>
-       		    	<prop key="perun.principal.extSourceName">cz.metacentrum.perun.core.impl.ExtSourceInternal</prop>
-       		    	<prop key="perun.principal.extSourceType">cz.metacentrum.perun.core.impl.ExtSourceInternal</prop>
-       		</props>	    
-       	</property> 
-    </bean>
+	<!-- Properties Bean -->
+	<bean id="dispatcherPropertiesBean" class="org.springframework.beans.factory.config.PropertiesFactoryBean">
+		<property name="locations">
+			<list>
+				<value>file:${perun.conf}perun-dispatcher.properties</value>
+				<value>file:${perun.conf.custom}perun-dispatcher.properties</value>
+			</list>
+		</property>
+		<property name="ignoreResourceNotFound">
+			<value>true</value>
+		</property>
+		<property name="properties">
+			<props>
+				<prop key="perun.principal.name">perunDispatcher</prop>
+				<prop key="perun.principal.extSourceName">INTERNAL</prop>
+				<prop key="perun.principal.extSourceType">cz.metacentrum.perun.core.impl.ExtSourceInternal</prop>
+			</props>
+		</property>
+	</bean>
 
-    <!-- BEGIN Quartz tasks -->
+	<!-- BEGIN Quartz tasks -->
 	<bean id="maintenanceJob" class="org.springframework.scheduling.quartz.JobDetailBean">
 		<property name="jobClass" value="cz.metacentrum.perun.dispatcher.job.MaintenanceJob" />
 		<property name="jobDataAsMap">
@@ -90,7 +90,7 @@ http://www.springframework.org/schema/task http://www.springframework.org/schema
 		</property>
 	</bean>
 	 -->
-	 
+
 	<bean id="maintenanceJobTrigger" class="org.springframework.scheduling.quartz.CronTriggerBean">
 		<property name="jobDetail" ref="maintenanceJob" />
 		<!-- Every 20 minutes -->
@@ -102,14 +102,14 @@ http://www.springframework.org/schema/task http://www.springframework.org/schema
 	<!--  
 	<bean id="checkInJobTrigger" class="org.springframework.scheduling.quartz.CronTriggerBean">
 		<property name="jobDetail" ref="checkInJob" /> -->
-		<!-- Every 3 minutes -->
+	<!-- Every 3 minutes -->
 	<!-- <property name="cronExpression" value="0 0/5 * * * ?" />-->
 	<!-- Every 10 seconds -->
-		<!-- property name="cronExpression" value="0/10 * * * * ?" /-->
+	<!-- property name="cronExpression" value="0/10 * * * * ?" /-->
 	<!--
 	</bean>
 	-->
-	
+
 	<bean class="org.springframework.scheduling.quartz.SchedulerFactoryBean">
 		<property name="triggers">
 			<list>
@@ -119,8 +119,8 @@ http://www.springframework.org/schema/task http://www.springframework.org/schema
 				<ref bean="propagationMaintainerJobTrigger" />
 			</list>
 		</property>
-	</bean>  
-	
+	</bean>
+
 	<bean id="propagationMaintainerJob" class="org.springframework.scheduling.quartz.JobDetailBean">
 		<property name="jobClass" value="cz.metacentrum.perun.dispatcher.job.PropagationMaintainerJob" />
 		<property name="jobDataAsMap">
@@ -131,10 +131,10 @@ http://www.springframework.org/schema/task http://www.springframework.org/schema
 	</bean>
 
 	<bean id="propagationMaintainerJobTrigger" class="org.springframework.scheduling.quartz.CronTriggerBean">
-		<property name="jobDetail" ref="propagationMaintainerJob" />	
-		<property name="cronExpression" value="45 0/4 * * * ?" />	
+		<property name="jobDetail" ref="propagationMaintainerJob" />
+		<property name="cronExpression" value="45 0/4 * * * ?" />
 	</bean>
-	
+
 	<bean id="processPoolJob" class="org.springframework.scheduling.quartz.JobDetailBean">
 		<property name="jobClass" value="cz.metacentrum.perun.dispatcher.job.ProcessPoolJob" />
 		<property name="jobDataAsMap">
@@ -143,11 +143,11 @@ http://www.springframework.org/schema/task http://www.springframework.org/schema
 			</map>
 		</property>
 	</bean>
-	
+
 	<bean id="processPoolJobTrigger" class="org.springframework.scheduling.quartz.CronTriggerBean">
 		<property name="jobDetail" ref="processPoolJob" />
 		<property name="cronExpression" value="0 0/2 * * * ?" />
 	</bean>
-	<!-- END Quartz tasks -->	
-           
+	<!-- END Quartz tasks -->
+
 </beans>
