+++ perun-dispatcher-new/src/test/resources/perun-dispatcher-applicationcontext-test.xml
@@ -12,67 +12,65 @@ http://www.springframework.org/schema/aop http://www.springframework.org/schema/
 http://www.springframework.org/schema/tx http://www.springframework.org/schema/tx/spring-tx.xsd
 http://www.springframework.org/schema/task http://www.springframework.org/schema/task/spring-task.xsd
 "
-		default-lazy-init="true"
->
+       default-lazy-init="true"
+		>
 
-    <!-- Enable @Transactional support -->
-    <tx:annotation-driven/>
-    
-    <!-- Enable @AspectJ support -->
-    <aop:aspectj-autoproxy/>
-    
-    <!--It basically says, wrap any interface in the service package that ends
+	<!-- Enable @Transactional support -->
+	<tx:annotation-driven/>
+
+	<!-- Enable @AspectJ support -->
+	<aop:aspectj-autoproxy/>
+
+	<!--It basically says, wrap any interface in the service package that ends
 	with "Manager" with a transaction. Because of this, it's important to
 	set your dependencies to interfaces, not to implementations.
-    <aop:config>
-        <!<aop:advisor id="managerTx" advice-ref="txAdvice" pointcut="execution(* *..service.*Manager.*(..))"/>   
-    </aop:config>-->
-    
-    <!-- Importing Perun Core... -->
-    <import resource="classpath:perun-beans.xml"/>
-    	
-    <tx:advice id="txAdvice">
-        <tx:attributes>
-            <tx:method name="get*" read-only="true"/>
-            <tx:method name="*"/>
-        </tx:attributes>
-    </tx:advice>
+	<aop:config>
+		<!<aop:advisor id="managerTx" advice-ref="txAdvice" pointcut="execution(* *..service.*Manager.*(..))"/>
+	</aop:config>-->
 
-    <!-- Scans for @Repository, @Service and @Component -->
-    <context:component-scan base-package="cz.metacentrum.perun.dispatcher"/>
-    <context:component-scan base-package="cz.metacentrum.perun.controller"/>
-    
+	<!-- Importing Perun Core... -->
+	<import resource="classpath:perun-beans.xml"/>
+
+	<tx:advice id="txAdvice">
+		<tx:attributes>
+			<tx:method name="get*" read-only="true"/>
+			<tx:method name="*"/>
+		</tx:attributes>
+	</tx:advice>
+
+	<!-- Scans for @Repository, @Service and @Component -->
+	<context:component-scan base-package="cz.metacentrum.perun.dispatcher"/>
+	<context:component-scan base-package="cz.metacentrum.perun.controller"/>
+
 	<!-- Properties Bean -->
 	<bean id="dispatcherPropertiesBean" class="org.springframework.beans.factory.config.PropertiesFactoryBean">
-        <property name="locations">
-        	<list>
-	           	<value>${perun.jdbc}</value>
-				<value>file:${perun.conf.custom}jdbc.properties</value>
-	            <value>file:${perun.conf}perun-dispatcher.properties</value>
-                <value>file:${perun.conf.custom}perun-dispatcher.properties</value>
-            </list>
-        </property>
-        <property name="ignoreResourceNotFound">
-            <value>true</value>
-        </property>
-       	<property name="properties">
-       		<props>
-       		    	<prop key="perun.principal.name">perunTests</prop>
-       		    	<prop key="perun.principal.extSourceName">cz.metacentrum.perun.core.impl.ExtSourceInternal</prop>
-       		    	<prop key="perun.principal.extSourceType">cz.metacentrum.perun.core.impl.ExtSourceInternal</prop>
-       		</props>	    
-       	</property> 
-   	</bean>
+		<property name="locations">
+			<list>
+				<value>file:${perun.conf}perun-dispatcher.properties</value>
+				<value>file:${perun.conf.custom}perun-dispatcher.properties</value>
+			</list>
+		</property>
+		<property name="ignoreResourceNotFound">
+			<value>true</value>
+		</property>
+		<property name="properties">
+			<props>
+				<prop key="perun.principal.name">perunDispatcher</prop>
+				<prop key="perun.principal.extSourceName">INTERNAL</prop>
+				<prop key="perun.principal.extSourceType">cz.metacentrum.perun.core.impl.ExtSourceInternal</prop>
+			</props>
+		</property>
+	</bean>
 
 	<bean id="taskExecutor" class="org.springframework.scheduling.concurrent.ThreadPoolTaskExecutor">
 		<property name="corePoolSize" value="10" />
-  		<property name="maxPoolSize" value="50" />
-  		<property name="queueCapacity" value="100" />
+		<property name="maxPoolSize" value="50" />
+		<property name="queueCapacity" value="100" />
 	</bean>
-		
+
 	<beans profile="integration,production">
-    
-    	<!-- BEGIN Quartz tasks -->	
+
+		<!-- BEGIN Quartz tasks -->
 		<bean id="maintenanceJob" class="org.springframework.scheduling.quartz.JobDetailBean">
 			<property name="jobClass" value="cz.metacentrum.perun.dispatcher.job.MaintenanceJob" />
 			<property name="jobDataAsMap">
@@ -81,7 +79,7 @@ http://www.springframework.org/schema/task http://www.springframework.org/schema
 				</map>
 			</property>
 		</bean>
-	
+
 		<!--  
 		<bean id="checkInJob" class="org.springframework.scheduling.quartz.JobDetailBean">
 			<property name="jobClass" value="cz.metacentrum.perun.dispatcher.job.CheckInJob" />
@@ -92,7 +90,7 @@ http://www.springframework.org/schema/task http://www.springframework.org/schema
 			</property>
 		</bean>
 		-->
-		
+
 		<bean id="maintenanceJobTrigger" class="org.springframework.scheduling.quartz.CronTriggerBean">
 			<property name="jobDetail" ref="maintenanceJob" />
 			<!-- Every 20 minutes -->
@@ -100,7 +98,7 @@ http://www.springframework.org/schema/task http://www.springframework.org/schema
 			<!-- Every 5 seconds -->
 			<property name="cronExpression" value="0/5 * * * * ?" />
 		</bean>
-	
+
 		<!--  
 		<bean id="checkInJobTrigger" class="org.springframework.scheduling.quartz.CronTriggerBean">
 			<property name="jobDetail" ref="checkInJob" />
@@ -110,7 +108,7 @@ http://www.springframework.org/schema/task http://www.springframework.org/schema
 			property name="cronExpression" value="0/10 * * * * ?" / 
 		</bean>
 		-->
-		
+
 		<bean class="org.springframework.scheduling.quartz.SchedulerFactoryBean">
 			<property name="triggers">
 				<list>
@@ -118,8 +116,8 @@ http://www.springframework.org/schema/task http://www.springframework.org/schema
 					<!--<ref bean="checkInJobTrigger" /> -->
 				</list>
 			</property>
-		</bean>  
-		<!-- END Quartz tasks -->	
-	</beans>          
+		</bean>
+		<!-- END Quartz tasks -->
+	</beans>
 
 </beans>
