+++ perun-rpc/pom.xml
@@ -112,17 +112,6 @@
 
 		</plugins>
 
-		<resources>
-			<resource>
-				<directory>src/main/resources</directory>
-				<filtering>true</filtering>
-				<!-- Exclude testing data -->
-				<excludes>
-					<exclude>test-*.txt</exclude>
-				</excludes>
-			</resource>
-		</resources>
-
 	</build>
 
 	<dependencies>
