+++ perun-rpc/src/main/webapp/WEB-INF/web.xml
@@ -16,7 +16,7 @@
 
     <context-param>
         <param-name>log4jConfigLocation</param-name>
-        <param-value>file://${perun.conf}log4j.properties</param-value>
+        <param-value>file://${perun.conf}log4j.xml</param-value>
     </context-param>
 
     <!-- !! MUST BE FIRST LISTENER IN THIS CONFIG FILE !! -->
@@ -42,7 +42,6 @@
             classpath:perun-notification-applicationcontext-scheduling.xml
             classpath:perun-dispatcher-applicationcontext-jdbc.xml
             classpath:perun-dispatcher-applicationcontext.xml
-            classpath:perun-rpc-dispatcher-applicationcontext.xml
         </param-value>
     </context-param>
